allow for NA values in existing
Benjamin Bright testing showed that the use of existing samples that are located in stratum where values are NA results in some algorithms failing.

Inserted ways to notify the user that samples (and how many) fall in NA areas and ensure sample_strat functions appropriately whether include = TRUE or not.